+++ packages/agent-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vibeflow/agent-sdk
 
+## 1.1.0
+
+### Minor Changes
+
+- Added agents
+
 ## 1.0.0
 
 ### Major Changes