+++ packages/replier/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @vibeflow/replier
 
+## 2.1.0
+
+### Minor Changes
+
+- Added agents
+
+### Patch Changes
+
+- Updated dependencies
+  - @vibeflow/agent-sdk@1.1.0
+  - @vibeflow/core@1.1.0
+  - @vibeflow/database@2.1.0
+
 ## 2.0.0
 
 ### Major Changes